Define pinnedServerThreadPool in RequestCpuAccounting so the base-installed harness compiles
build-explore-jmh installs the base-owned ExplorerMemoryBenchmark.kt over the
candidate tree but keeps the candidate RequestCpuAccounting.kt. With the pool
helper defined in ExplorerMemoryBenchmark.kt, the candidate contract in
RequestCpuAccounting.kt referenced a symbol the base harness does not have, so
:explore:compileJmhKotlin failed with "Unresolved reference 'pinnedServerThreadPool'".

Move the helper (and its constants) into RequestCpuAccounting.kt, which the
candidate keeps; the benchmark harness calls it unqualified within the same
package. The base harness has no reference to it and compiles unchanged.

Verified by installing origin/main's ExplorerMemoryBenchmark.kt over the
candidate RequestCpuAccounting.kt and compiling :explore:compileJmhKotlin
(the build-explore-jmh scenario) -- BUILD SUCCESSFUL -- and the contract still
passes locally (present-then-gone fails closed, pinned pool holds).
